pyproject.toml
[project]
name = "data-management-plan-tooling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "pywin32>=308",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.4",
]
